add notice about missing nameSpaces field in CBOR for ISO 23220-4
