+++ document-manager/src/main/java/eu/europa/ec/eudi/wallet/document/DocumentManager.kt
@@ -83,7 +83,7 @@ interface DocumentManager {
      *
      * ```cddl
      * IssuerSigned = {
-     *   "nameSpaces" : IssuerNameSpaces, ; Returned data elements
+     *   ?"nameSpaces" : IssuerNameSpaces, ; Returned data elements
      *   "issuerAuth" : IssuerAuth ; Contains the mobile security object (MSO) for issuer data authentication
      * }
      * IssuerNameSpaces = { ; Returned data elements for each namespace
@@ -98,6 +98,8 @@ interface DocumentManager {
      * }
      * IssuerAuth = COSE_Sign1 ; The payload is MobileSecurityObjectBytes
      * ```
+     *
+     * **Important** Currently `nameSpaces` field should exist and must not be empty.
      *
      * The document is added in the storage and can be retrieved using the
      * [DocumentManager::getDocumentById] method.
